remove todo methods
