+++ generators/liquibase/generator.mts
@@ -28,7 +28,7 @@ import {
   prepareRelationshipForLiquibase,
   liquibaseComment,
 } from './support/index.mjs';
-import { prepareEntity as prepareEntityForServer } from '../server/support/index.mjs';
+import { getFKConstraintName, getUXConstraintName, prepareEntity as prepareEntityForServer } from '../server/support/index.mjs';
 import {
   prepareEntityPrimaryKeyForTemplates,
   prepareRelationship,
@@ -663,5 +663,35 @@ export default class LiquibaseGenerator extends BaseEntityChangesGenerator {
       return this._addUpdateFilesReferences({ entity, databaseChangelog, changelogData, source });
     }
     return undefined;
+  }
+
+  /**
+   * @private
+   * get a foreign key constraint name for tables in JHipster preferred style.
+   *
+   * @param {string} entityName - name of the entity
+   * @param {string} relationshipName - name of the related entity
+   * @param {string} prodDatabaseType - database type
+   * @param {boolean} noSnakeCase - do not convert names to snakecase
+   */
+  getFKConstraintName(entityName, relationshipName, prodDatabaseType, noSnakeCase) {
+    const result = getFKConstraintName(entityName, relationshipName, { prodDatabaseType, noSnakeCase });
+    (this as any).validateResult(result);
+    return result.value;
+  }
+
+  /**
+   * @private
+   * get a unique constraint name for tables in JHipster preferred style.
+   *
+   * @param {string} entityName - name of the entity
+   * @param {string} columnName - name of the column
+   * @param {string} prodDatabaseType - database type
+   * @param {boolean} noSnakeCase - do not convert names to snakecase
+   */
+  getUXConstraintName(entityName, columnName, prodDatabaseType, noSnakeCase) {
+    const result = getUXConstraintName(entityName, columnName, { prodDatabaseType, noSnakeCase });
+    (this as any).validateResult(result);
+    return result.value;
   }
 }
